separeted the mail sending file and tested the momentjs lib
