+++ server/package.json
@@ -14,6 +14,7 @@
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
+    "moment": "^2.24.0",
     "nodemailer": "^6.4.4",
     "nodemon": "^2.0.2",
     "pg": "^7.18.2",
